New Notebook
New notebook , holds all data integration and prepping the dataset and feature engineering. Also has some EDA (data analysis) which particularly looks at the correlation of certain features (statistics) and winning games in the tourney, and also looks at the distribution of features in the dataset. Then there is training and testing the model in conjunction with feature selection/reduction methods.... What's left to be done is finalizing feature selection, model selection and hyper-parameter tuning.